+++ container.go
@@ -2,18 +2,28 @@ package epub
 
 import (
 	"encoding/xml"
+	"io"
 )
 
 const (
 	containerPath = "META-INF/container.xml"
 )
 
-type containerXML struct {
-	XmlName   xml.Name `xml:"urn:oasis:names:tc:opendocument:xmlns:container container"`
+type container struct {
+	XMLName   xml.Name `xml:"urn:oasis:names:tc:opendocument:xmlns:container container"`
 	Rootfiles rootfile `xml:"rootfiles>rootfile"`
 }
 
 type rootfile struct {
-	XmlName  xml.Name `xml:"rootfile"`
+	XMLName  xml.Name `xml:"rootfile"`
 	FullPath string   `xml:"full-path,attr"`
+}
+
+func newContainer(r io.Reader) (*container, error) {
+	c := &container{}
+	if err := decodeXML(r, &c); err != nil {
+		return nil, err
+	}
+
+	return c, nil
 }