+++ epub_test.go
@@ -1,9 +1,9 @@
 package epub
 
 import (
-	"testing"
-	"path/filepath"
 	"encoding/json"
+	"path/filepath"
+	"testing"
 
 	"github.com/pirmd/verify"
 )
@@ -12,13 +12,13 @@ const (
 	testdataPath = "./testdata"
 )
 
-func TestGetMetadata(t *testing.T) {
+func TestGetMetadataFromFile(t *testing.T) {
 	testCases, err := filepath.Glob(filepath.Join(testdataPath, "*.epub"))
 	if err != nil {
 		t.Fatalf("cannot read test data in %s:%v", testdataPath, err)
 	}
 
-	out := []*Metadata{}
+	out := []*MetaInformation{}
 	for _, tc := range testCases {
 		m, err := GetMetadataFromFile(tc)
 		if err != nil {
@@ -32,7 +32,32 @@ func TestGetMetadata(t *testing.T) {
 		t.Fatalf("Fail to marshal test output to json: %v", err)
 	}
 
-    if failure := verify.MatchGolden(t.Name(), string(got)); failure != nil {
-        t.Fatalf("Metadata is not as expected:\n%v", failure)
-    }
+	if failure := verify.MatchGolden(t.Name(), string(got)); failure != nil {
+		t.Fatalf("Metadata is not as expected:\n%v", failure)
+	}
+}
+
+func TestGetPackageFromFile(t *testing.T) {
+	testCases, err := filepath.Glob(filepath.Join(testdataPath, "*.epub"))
+	if err != nil {
+		t.Fatalf("cannot read test data in %s:%v", testdataPath, err)
+	}
+
+	out := []*PackageDocument{}
+	for _, tc := range testCases {
+		opf, err := GetPackageFromFile(tc)
+		if err != nil {
+			t.Errorf("Fail to get package for %s: %v", tc, err)
+		}
+		out = append(out, opf)
+	}
+
+	got, err := json.MarshalIndent(out, "", "  ")
+	if err != nil {
+		t.Fatalf("Fail to marshal test output to json: %v", err)
+	}
+
+	if failure := verify.MatchGolden(t.Name(), string(got)); failure != nil {
+		t.Fatalf("Package Document is not as expected:\n%v", failure)
+	}
 }