+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## [0.2.0]
+## [0.2.0] - 2022-06-23
 - expose PackageDocument struct and add functions to get it from an epub.
 - add compliance to EPUB32 specifications https://www.w3.org/publishing/epub32/epub-packages.html.
 - add helpers (WalkXxX) to access EPUB publication resources.